Revert "Fixes on Role Manager"
This reverts commit 2a3c81d93a45a3c8143de07ad503ad65b88f13e2.

+++ Controllers/UserManagementController.cs
@@ -141,10 +141,7 @@ namespace NetCoreBootstrap.Controllers
                 isCurrentRole = userRoles.Contains(role.Key);
                 if(isCurrentRole)
                 {
-                    if(!role.Value)
-                    {
-                        await UserRepository.RemoveRoleFromUser(user, role.Key);
-                    }
+                    if(!role.Value) await UserRepository.RemoveRoleFromUser(user, role.Key);
                 }
                 else if(role.Value) await UserRepository.AddRoleToUser(user, role.Key);
             }
